Add files via upload
This version takes a different approach -- rather than filling in single gaps in rows, cols, and blocks, it eliminates candidates in each unknown value and fills them in when an unknown value has only a single candidate left. This allows it to use the combined information of an unknown value's known row, column, and block values. 

The next step for v3 is to enable an unknown value to use information from the candidates of other unknown values in its row, col, and block (e.g., if only one unknown value in a block has a 3 as a candidate and the block doesn't already have a 3, that unknown value must be a 3).